cleanup modal - part I
